Add ability to export a shader AOV to the compositing layers
Unfortunately, Blender doens't support an easy way of rendering an AOV based on shaders. So, this script loops through the materials in the scene and adds an AOV Output shader to the node tree. It connects to the AOV output the same things that were plugged into the given shader, with support for mixing and adding the value in the same way that the shaders were mixed or added.